Fixed the post-build copies to take care of the name decorating needed by the installer.  We don't name decorate the binaries at build time because that gives VS heartburn.  Updated the readme.  Switched to level 4 warnings and made warnings errors.

+++ trunk/TWAIN-Samples/Twain_App_sample01/visual_studio_mfc/TW_Enum_Dlg.cpp
@@ -49,7 +49,7 @@ IMPLEMENT_DYNAMIC(CTW_Enum_Dlg, CDialog)
 CTW_Enum_Dlg::CTW_Enum_Dlg(CWnd* pParent /*=NULL*/)
   : CDialog(CTW_Enum_Dlg::IDD, pParent)
   ,m_pCap(NULL)
-  ,m_SelectionData(-1)
+  ,m_SelectionData((DWORD)-1)
   ,m_SelectionString(_T(""))
 {
 }
